Add confirm dialog modal to client

+++ client/src/app/messages/messages.component.ts
@@ -1,7 +1,12 @@
-// libs
+// Libs
 import { Component, OnInit } from '@angular/core';
+
+// Models
 import { Message } from '../models/message';
 import { Pagination } from '../models/pagination';
+
+// Services
+import { ConfirmService } from '../services/confirm.service';
 import { MessageService } from '../services/message.service';
 
 @Component({
@@ -17,7 +22,10 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) {}
+  constructor(
+    private messageService: MessageService,
+    private confirmService: ConfirmService
+  ) {}
 
   ngOnInit(): void {
     this.loadMessages();
@@ -35,12 +43,18 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number): void {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(
-        this.messages.findIndex((message) => message.id === id),
-        1
-      );
-    });
+    this.confirmService
+      .confirm('Confirm delete message', 'This cannot be undone')
+      .subscribe((result) => {
+        if (result) {
+          this.messageService.deleteMessage(id).subscribe(() => {
+            this.messages.splice(
+              this.messages.findIndex((message) => message.id === id),
+              1
+            );
+          });
+        }
+      });
   }
 
   pageChanged(event: any): void {
